Added schedule save success message. Added update of the minimum staffing when a position is selected.

+++ project/LaborGuru_SPM/trunk/src/main/java/com/laborguru/action/schedule/AddShiftBaseAction.java
@@ -631,7 +631,7 @@ public abstract class AddShiftBaseAction extends SpmAction {
 		int rowCount = 1;
 		
 		for(ScheduleRow aRow : schedule) {
-			if(aRow.getEmployeeId() == null || aRow.getEmployeeId() <= 0) {
+			if(aRow.getEmployeeId() == null || aRow.getEmployeeId().intValue() <= 0) {
 				addActionError(getText("error.schedule.addshift.employee_missing", new String[]{String.valueOf(rowCount)}));
 			}
 			rowCount++;
